requirements_test.txt
# Requirements for LoRa Dongle testing
pyserial>=3.5
